Fixed more windows warnings
- Fixed uninitialized variables in default cases
- Fixed truncation of values to float with static_cast

Reviewed-on: https://swiftshader-review.googlesource.com/5682
Tested-by: Alexis Hétu <sugoi@google.com>

+++ src/OpenGL/libGLESv2/utilities.cpp
@@ -1180,7 +1180,7 @@ namespace es2
 		return false;
 	}
 
-	std::string ParseUniformName(const std::string &name, size_t *outSubscript)
+	std::string ParseUniformName(const std::string &name, int *outSubscript)
 	{
 		// Strip any trailing array operator and retrieve the subscript
 		size_t open = name.find_last_of('[');
